requirements.txt
pyspark
pyiceberg[spark]
python-dotenv
kafka-python